feat(planning): SMS delivery status on confirmation badges (Wave 7 slice 7)
Close the loop opened in slice 6: surface Twilio's delivery status so the planner
sees whether an order-confirmation SMS actually reached the customer, not just
that it was sent.

- Webhook: app/api/twilio/sms-status additively updates ops_work_order_confirmations
  by provider_message_id (best-effort, alongside the existing planning_project_meta
  update — a MessageSid lives in at most one table). The new planning's SMS already
  routes here via TWILIO_STATUS_CALLBACK_URL, so no send-side change was needed.
- Domain: ConfirmationSummary gains sms_status (latest Twilio status per work order)
  + a pure describeSmsStatus() → { label, tone } (delivered=ok, failed/undelivered=
  fail, in-flight=pending). summarizeConfirmations captures the newest sms row's
  status; confirmationsByWorkOrder selects it.
- UI: ConfirmationBadge turns rose with an alert icon + "ej levererat" when the SMS
  failed; month-card check shows a rose "!" instead of the green tick. Tooltip
  includes the delivery label.
- Module split: the Node-only send half (Resend/Twilio) moved to
  lib/domains/planning/confirmationsSend.ts so board cards can import the pure
  describeSmsStatus from ./confirmations without pulling fs/net/tls into the client
  bundle (fixes a webpack 'Can't resolve fs' regression).
- Tests: describeSmsStatus mapping + summary captures latest sms_status.

No new migration (status column already exists from slice 6).
type-check clean · 401 tests green · build OK.
